Minor edit to md2html.sh
Hoping that the reason that stumash/github.io/Algorithms_Notes is
not rendering Mathjax is that the link was http and not https.  So,
changed the link to https.  Also changed the name of the function
pandochtmlmath defined in md2html.sh to pandochtmlmath_local since
that function was already defined on my machine and I didn't want
namespace conflicts.